+++ NEWS.md
@@ -1,3 +1,7 @@
+# 0.4.22
+  * Reorder checks: move check package near the top
+  * Don't enforce version changes on RELEASE: FALSE
+
 # 0.4.21
   * Automatically update config files